api changes
move logout to auth routes
