+++ backend/src/routes/auth.route.js
@@ -1,9 +1,10 @@
 const express = require('express')
-const { register, login } = require('../controllers/auth.controller')
+const { register, login, logout } = require('../controllers/auth.controller')
 
 const router = express.Router()
 
 router.post('/register', register)
 router.post('/login', login)
+router.post('/logout', logout)
 
 module.exports = router
