+++ backend/src/routes/user.route.js
@@ -3,7 +3,6 @@ const {
   deleteUser,
   getUser,
   getUsers,
-  logout,
   updateUser,
   updateUserAdmin
 } = require('../controllers/user.controller')
@@ -11,11 +10,10 @@ const { verifyToken, verifyIsAdmin } = require('../utils/auth')
 
 const router = express.Router()
 
-router.put('/update/:userId', verifyToken, updateUser)
-router.put('/admin/:userId', verifyToken, verifyIsAdmin, updateUserAdmin)
-router.delete('/delete/:userId', verifyToken, deleteUser)
-router.post('/logout', logout)
-router.get('/getusers', verifyToken, verifyIsAdmin, getUsers)
+router.get('/', verifyToken, verifyIsAdmin, getUsers)
+router.patch('/:userId', verifyToken, updateUser)
 router.get('/:userId', verifyToken, getUser)
+router.delete('/:userId', verifyToken, deleteUser)
+router.put('/admin/:userId', verifyToken, verifyIsAdmin, updateUserAdmin)
 
 module.exports = router
